BD-RL-003-fullstack_meeting_app/requirements.txt
# Python dependencies for TASK_005 benchmark testing
# These are required to run the test suite

# Test framework
pytest==9.0.2
pytest-asyncio==0.24.0
pytest-anyio==4.12.0

# HTTP client for API testing
httpx==0.28.1

# Database and ORM
sqlalchemy==2.0.36
psycopg[binary]==3.2.3

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Date/time utilities
python-dateutil==2.9.0.post0

# Backend application dependencies (for testing)
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3

